+++ wallnut.cpp
@@ -2,7 +2,7 @@
 
 wallnut::wallnut()
 {
-    hp=3000;
+    hp=3000.0;
     setMovie(":/new/prefix1/WallNut.gif");
 }
 void wallnut::advance(int phase)
@@ -10,17 +10,17 @@ void wallnut::advance(int phase)
     if(!phase)
         return;
 
-    if(hp<=0)
+    if((int)hp<=0)
     {
         delete this;
         return;
     }
-    else if(hp<1000)
+    else if((int)hp<1000)
     {
         setMovie(":/new/prefix1/WallNut2.gif");
         return;
     }
-    else if(hp<2000)
+    else if((int)hp<2000)
     {
         setMovie(":/new/prefix1/WallNut1.gif");
         return;
